Updated styling
