+++ src/components/christmasHomePageDecoration/modal/greetingModal/background/Background.tsx
