+++ src/components/christmasHomePageDecoration/modal/greetingModal/background/BgDesktop.tsx
