+++ src/components/christmasHomePageDecoration/modal/greetingModal/background/BgMobile.tsx
