+++ src/components/christmasHomePageDecoration/modal/greetingModal/background/BgPc.tsx
