+++ src/components/christmasHomePageDecoration/modal/greetingModal/background/BgTablet.tsx
