+++ src/components/shared/Icons/christmas/IconBall.tsx
@@ -12,11 +12,11 @@ export const IconBall = ({ className }: { className: string }) => {
         >
             <path
                 d="M24.9386 87.6389C24.0484 87.6389 23.1781 87.3749 22.4379 86.8803C21.6977 86.3858 21.1208 85.6828 20.7801 84.8603C20.4395 84.0378 20.3503 83.1328 20.524 82.2597C20.6977 81.3866 21.1264 80.5846 21.7559 79.9551C22.3853 79.3256 23.1873 78.8969 24.0605 78.7232C24.9336 78.5495 25.8386 78.6387 26.6611 78.9793C27.4835 79.32 28.1865 79.8969 28.6811 80.6371C29.1757 81.3773 29.4397 82.2476 29.4397 83.1378C29.4397 84.3316 28.9655 85.4765 28.1213 86.3206C27.2772 87.1647 26.1324 87.6389 24.9386 87.6389ZM24.9386 79.4829C24.2264 79.4829 23.5302 79.6941 22.9381 80.0898C22.3459 80.4855 21.8843 81.0478 21.6118 81.7058C21.3392 82.3638 21.268 83.0878 21.4069 83.7863C21.5458 84.4848 21.8888 85.1264 22.3924 85.63C22.896 86.1336 23.5376 86.4766 24.2361 86.6155C24.9346 86.7544 25.6586 86.6831 26.3166 86.4106C26.9745 86.138 27.5369 85.6765 27.9326 85.0843C28.3283 84.4922 28.5395 83.796 28.5395 83.0838C28.5395 82.1288 28.1601 81.2129 27.4848 80.5376C26.8095 79.8623 25.8936 79.4829 24.9386 79.4829Z"
-                fill="#A258FF"
+                fill="currentColor"
             />
             <path
                 d="M25.3846 -30H24.4844V79.0346H25.3846V-30Z"
-                fill="#A258FF"
+                fill="currentColor"
             />
             <path
                 d="M49.8881 108.056C49.8807 110.247 49.5779 112.427 48.9879 114.537L32.5859 109.757L48.6909 100.539C49.4719 102.969 49.8756 105.504 49.8881 108.056Z"
